remake AddLecture api for new structure of project
