+++ speakers/authapp/views.py
@@ -3,11 +3,7 @@ from rest_framework import permissions
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from django.contrib.auth.decorators import login_required
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.core.exceptions import ObjectDoesNotExist
-
-from .models import Person
+from workroomsapp.models import Person
 
 from .serializers import (
     UserProfileCreateSerializer,
@@ -49,33 +45,6 @@ class UserProfileLogoutView(APIView):
             status=201
         )
 
-
-class AddLecture(APIView):
-    def post(self, request):
-        if (request.user.is_authenticated or request.user.is_staff):
-            try:
-                user = Person.objects.get(person_userProfileId = request.user.pk)
-            except ObjectDoesNotExist:
-                return Response(
-                    data={"status":"error","description": "NoProfile","user_msg":"Необходимо заполнить профиль"},
-                    status=500
-                )
-            if user.isLecturer:
-                print(user)
-                return Response(
-                    data={"status":"ok"},
-                    status=200
-                )
-            else:
-                return Response(
-                    data={"status":"error","description": "WrongAuthorization","user_msg":"Только лекторы могут добавлять лекции"},
-                    status=403
-                )
-        else:
-            return Response(
-                data={"status":"error","description": "Unauthorized","user_msg":"Требуется авторизация"},
-                status=401
-            )
 # --------------------Временные представления для разработки-----------------------
